+++ aiohttp_things/__init__.py
@@ -1,4 +1,4 @@
 from aiohttp_things import views
 
-__version__ = '0.3.1'
+__version__ = '0.4.0'
 __all__ = ['views']