+++ example.py
@@ -1,7 +1,7 @@
 import json
 import uuid
 from aiohttp import web
-from aiohttp_things.views import ContextMixin, PrimaryKeyMixin
+from aiohttp_things.views import ItemMixin, JSONMixin
 
 
 def safe_json_value(value):
@@ -12,12 +12,11 @@ def safe_json_value(value):
         return str(value)
 
 
-class Base(web.View, ContextMixin, PrimaryKeyMixin):
+class Base(web.View, ItemMixin, JSONMixin):
     async def get(self):
-        pk_type = type(self.pk)
-        self.context['Type of primary key'] = safe_json_value(pk_type)
+        self.context['Type of primary key'] = safe_json_value(type(self.pk))
         self.context['Value of primary key'] = safe_json_value(self.pk)
-        return web.json_response(self.context)
+        return await self.finalize_response()
 
 
 class IntegerExample(Base):